Split Event Repository
