+++ src/Sekiban.Core/Documents/IEventPersistentRepository.cs
@@ -1,0 +1,3 @@
+namespace Sekiban.Core.Documents;
+
+public interface IEventPersistentRepository : IEventRepository;
