+++ src/Sekiban.Core/Documents/IEventPersistentWriter.cs
@@ -1,0 +1,3 @@
+namespace Sekiban.Core.Documents;
+
+public interface IEventPersistentWriter : IEventWriter;
