+++ src/Sekiban.Core/Documents/IEventTemporaryRepository.cs
@@ -1,0 +1,3 @@
+namespace Sekiban.Core.Documents;
+
+public interface IEventTemporaryRepository : IEventRepository;
