+++ src/Sekiban.Core/Documents/IEventTemporaryWriter.cs
@@ -1,0 +1,3 @@
+namespace Sekiban.Core.Documents;
+
+public interface IEventTemporaryWriter : IEventWriter;
