+++ src/Sekiban.Core/Documents/IEventWriter.cs
@@ -1,0 +1,8 @@
+using Sekiban.Core.Documents.Pools;
+using Sekiban.Core.Events;
+namespace Sekiban.Core.Documents;
+
+public interface IEventWriter
+{
+    Task SaveEvents<TEvent>(IEnumerable<TEvent> events, IWriteDocumentStream writeDocumentStream) where TEvent : IEvent;
+}
