+++ src/Sekiban.Core/Documents/Pools/IAggregatesStream.cs
@@ -1,6 +1,5 @@
 using ResultBoxes;
 using Sekiban.Core.Aggregate;
-using Sekiban.Core.Types;
 namespace Sekiban.Core.Documents.Pools;
 
 public interface IAggregatesStream
@@ -16,21 +15,4 @@ public interface IAggregatesStream
                     : new ApplicationException("Stream Names is not set"))
             .Conveyor(_ => GetStreamNames()[0].ToResultBox());
 }
-public interface IWriteDocumentStream
-{
-    public AggregateContainerGroup GetAggregateContainerGroup();
-}
-public record AggregateWriteStream(Type AggregatePayloadType) : IWriteDocumentStream
-{
-    public AggregateContainerGroup GetAggregateContainerGroup() =>
-        AggregateContainerGroupAttribute.FindAggregateContainerGroup(GetOriginal());
-    public Type GetOriginal() => AggregatePayloadType.IsAggregatePayloadType()
-        ? AggregatePayloadType.GetBaseAggregatePayloadTypeFromAggregate()
-        : AggregatePayloadType;
-}
 //TBD
-public record PureAggregateProjectionWriteStream(Type PureAggregateProjectionType) : IWriteDocumentStream
-{
-    public AggregateContainerGroup GetAggregateContainerGroup() =>
-        AggregateContainerGroupAttribute.FindAggregateContainerGroup(PureAggregateProjectionType);
-}
