+++ src/Sekiban.Core/Documents/Pools/IPooledDocumentWriter.cs
@@ -1,0 +1,6 @@
+namespace Sekiban.Core.Documents.Pools;
+
+public interface IPooledDocumentWriter : IDocumentWriter
+{
+    public int GetPoolOrderIndex();
+}
