+++ src/Sekiban.Core/Documents/Pools/IPooledEventRepository.cs
@@ -1,0 +1,6 @@
+namespace Sekiban.Core.Documents.Pools;
+
+public interface IPooledEventRepository : IEventRepository
+{
+    public int GetPoolOrderIndex();
+}
