+++ src/Sekiban.Core/Documents/Pools/IWriteDocumentStream.cs
@@ -1,0 +1,7 @@
+using Sekiban.Core.Aggregate;
+namespace Sekiban.Core.Documents.Pools;
+
+public interface IWriteDocumentStream
+{
+    public AggregateContainerGroup GetAggregateContainerGroup();
+}
