+++ src/Sekiban.Core/Documents/Pools/PureAggregateProjectionWriteStream.cs
@@ -1,0 +1,8 @@
+using Sekiban.Core.Aggregate;
+namespace Sekiban.Core.Documents.Pools;
+
+public record PureAggregateProjectionWriteStream(Type PureAggregateProjectionType) : IWriteDocumentStream
+{
+    public AggregateContainerGroup GetAggregateContainerGroup() =>
+        AggregateContainerGroupAttribute.FindAggregateContainerGroup(PureAggregateProjectionType);
+}
